+++ api/helpers/meta/change.js
@@ -181,6 +181,11 @@ module.exports = {
         return moment(value).isValid()
       }
     },
+    delaySystem: {
+      type: 'number',
+      allowNull: true,
+      description: 'Number of seconds currently on the delay system. Null means delay system might be offline or delay system is in bypass mode.'
+    },
     webchat: {
       type: 'boolean',
       description: 'Set to false to restrict the ability to send chat messages through the website'