+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - "originalStart" and "originalEnd" columns for Calendar; these contain the original start and end times for events for comparison when times are changed.
  - "delaySystem" option to Hosts (can only be set to one host). Used to specify which host is responsible for / connected to the delay system.
  - "EAS" option to Hosts (can only be set to one host). Used to specify which host is responsible for / connected to the Emergency Alery System.
+ - Meta.delaySystem; number of seconds currently on ther delay. Could be null if delay is offline or in bypass mode.
 
 ### Changed
  - Meta.queueFinish time will now include music tracks / openers when starting a broadcast to avoid confusion on the DJ Controls queue countdown. State will still switch to "*_on" after noMeta tracks have been played to trigger the recorder in DJ Controls.